refactor: extract transactionThreadReportID computation into seperate hook
