+++ src/hooks/useTransactionThreadReport.ts
@@ -1,11 +1,7 @@
 import getNonEmptyStringOnyxID from '@libs/getNonEmptyStringOnyxID';
-import {getAllNonDeletedTransactions} from '@libs/MoneyRequestReportUtils';
-import {getFilteredReportActionsForReportView, getOneTransactionThreadReportID, isSentMoneyReportAction} from '@libs/ReportActionsUtils';
 import ONYXKEYS from '@src/ONYXKEYS';
-import useNetwork from './useNetwork';
 import useOnyx from './useOnyx';
-import usePaginatedReportActions from './usePaginatedReportActions';
-import useReportTransactionsCollection from './useReportTransactionsCollection';
+import useTransactionThreadReportID from './useTransactionThreadReportID';
 
 /**
  * Derives the single-transaction thread report ID and report for a money request report.
@@ -14,27 +10,12 @@ import useReportTransactionsCollection from './useReportTransactionsCollection';
  * whether a report has a single transaction thread (and access its data).
  */
 function useTransactionThreadReport(reportID: string | undefined) {
-    const {isOffline} = useNetwork();
+    const {transactionThreadReportID, reportActions} = useTransactionThreadReportID(reportID);
 
-    const [moneyRequestReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${reportID}`);
-    const [chatReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${getNonEmptyStringOnyxID(moneyRequestReport?.chatReportID)}`);
-
-    const {reportActions: unfilteredReportActions} = usePaginatedReportActions(moneyRequestReport?.reportID);
-    const reportActions = getFilteredReportActionsForReportView(unfilteredReportActions);
-
-    const allReportTransactions = useReportTransactionsCollection(reportID);
-    const nonDeletedTransactions = getAllNonDeletedTransactions(allReportTransactions, reportActions, isOffline, true);
-    const visibleTransactions = nonDeletedTransactions?.filter((t) => isOffline || t.pendingAction !== 'delete');
-    const reportTransactionIDs = visibleTransactions?.map((t) => t.transactionID);
-
-    const transactionThreadReportID = getOneTransactionThreadReportID(moneyRequestReport, chatReport, reportActions ?? [], isOffline, reportTransactionIDs);
-    const isSentMoneyReport = reportActions.some((action) => isSentMoneyReportAction(action));
-    const effectiveTransactionThreadReportID = isSentMoneyReport ? undefined : transactionThreadReportID;
     const [transactionThreadReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${getNonEmptyStringOnyxID(transactionThreadReportID)}`);
 
     return {
         transactionThreadReportID,
-        effectiveTransactionThreadReportID,
         transactionThreadReport,
         reportActions,
     };
